+++ src/app/app.module.ts
@@ -21,10 +21,10 @@ import { TabsPage } from '../pages/tabs/tabs';
 import { TutorialPage } from '../pages/tutorial/tutorial';
 import { WelcomePage } from '../pages/welcome/welcome';
 
-import { Api } from '../providers/api';
+import { Api } from '../providers/providers';
 import { Items } from '../mocks/providers/items';
-import { Settings } from '../providers/settings';
-import { User } from '../providers/user';
+import { Settings } from '../providers/providers';
+import { User } from '../providers/providers';
 
 import { Camera } from '@ionic-native/camera';
 import { GoogleMaps } from '@ionic-native/google-maps';
