+++ src/pages/login/login.ts
@@ -3,7 +3,7 @@ import { NavController, ToastController } from 'ionic-angular';
 
 import { MainPage } from '../../pages/pages';
 
-import { User } from '../../providers/user';
+import { User } from '../../providers/providers';
 
 import { TranslateService } from '@ngx-translate/core';
