+++ src/pages/signup/signup.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { NavController, ToastController } from 'ionic-angular';
 
 import { MainPage } from '../../pages/pages';
-import { User } from '../../providers/user';
+import { User } from '../../providers/providers';
 
 import { TranslateService } from '@ngx-translate/core';
