+++ src/providers/items/items.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core';
 import { Http } from '@angular/http';
 import 'rxjs/add/operator/map';
 
-import { Api } from './api';
+import { Api } from '../api/api';
 
-import { Item } from '../models/item';
+import { Item } from '../../models/item';
 
 @Injectable()
 export class Items {
