+++ src/providers/user/user.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Http } from '@angular/http';
-import { Api } from './api';
+import { Api } from '../api/api';
+
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/toPromise';
